fix: remove unnecessary type cast in engine.ts tick loop
Agent-Logs-Url: https://github.com/A13Xg/Gray-Protocol/sessions/13a824f7-c5e3-4bb4-9054-84100246d46d

+++ src/core/engine.ts
@@ -77,7 +77,7 @@ export function tick(gs: GameState, deltaMs: number): void {
     // Apply costs/consumption
     for (const [key, amount] of Object.entries(delta.costs)) {
       const rk = key as ResourceKey;
-      const cost = new Decimal(amount as Decimal);
+      const cost = new Decimal(amount);
       if (cost.gt(0)) {
         if (gs.resources[rk].lt(cost)) {
           // Can't afford consumption, deactivate
